ubah tipe data

+++ src/main/java/id/ac/tazkia/akademik/aplikasiakademik/entity/Karyawan.java
@@ -41,7 +41,7 @@ public class Karyawan {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate tanggalLahir;
 
-    private BigInteger idAbsen;
+    private Integer idAbsen;
 
 
 }
